+++ .travis.yml
@@ -6,4 +6,4 @@ install:
   - pip install -r requirements.txt
 
 script:
-  - pylint --rcfile=.pylintrc web_pages web_page_elements functional models
+  - pylint --rcfile=.pylintrc pages elements functional models
