--- models/page_objects/base_page.py
@@ -1,6 +1,0 @@
-from selenium import webdriver
-
-
-class BasePage:
-    def __init__(self, driver=webdriver.Chrome()):
-        self.driver = driver
