+++ pages/mainpage.py
@@ -1,9 +1,7 @@
-from selenium import webdriver
-from models.driver import MyDriver
-from models.loginpage import LoginPage
+from pages.base import BasePage
 
 
-class MainPage(MyDriver):
+class MainPage(BasePage):
 
     def select_mac_product(self):
         """Make webdriver click Mac product."""
@@ -14,17 +12,3 @@ class MainPage(MyDriver):
         """Make webdriver add Mac product to Cart."""
         self.driver.find_element_by_xpath('//*[@id="content"]/div[2]/div/div/div[1]/a/img').click()
         self.driver.find_element_by_id("button-cart").click()
-
-# driver = webdriver.Chrome(executable_path=r'../webdrivers/chromedriver.exe')
-# base_url = "https://taqc296opencart.000webhostapp.com"
-# driver.get(base_url)
-#
-# page = LoginPage(driver)
-# page.go_to_login()
-# page.input_email()
-# page.input_password()
-# page.login()
-#
-# page = MainPage(driver)
-# page.select_mac_product()
-# page.add_mac_to_cart()
