+++ requirements.txt
@@ -1,14 +1,21 @@
 astroid==1.6.2
 attrs==17.4.0
 colorama==0.3.9
+enum34==1.1.6
 isort==4.3.4
+junit2html==20
 lazy-object-proxy==1.3.1
+lxml==4.2.0
 mccabe==0.6.1
-more-itertools==4.1.0
+mock==2.0.0
+namedlist==1.7
+pbr==3.1.1
 pluggy==0.6.0
-py==1.5.3
+py==1.5.2
 pylint==1.8.3
-pytest==3.5.0
+pytest==3.4.2
+pytest-allure-adaptor==1.7.10
 selenium==3.11.0
 six==1.11.0
+SQLAlchemy==1.2.5
 wrapt==1.10.11
